Fix quota map (#64)
* Fix the size for the disk quota info

In quota_maps, it stores both size and limit in it.
In function calculate_table_disk_usage, we only use
the "increment" part to update the size in quota_map.

So, if an entry in quota_map is vacuumed last time
and its size is some value, such as xx, but as it
has been vacuumed, then in the next loop its size
will be 0 or 0 + increment. 

* Fix SEGCOUNT
Before this commit, SEGCOUNT is 0 when the is_init 
is true, fix it.