Preserve AO storage options during internal CTAS
This change is an extension of the PR #13246.
The commit c121bcd83c9 fixes the bug for
ALTER TABLE EXPAND TABLE and ALTER TABLE SET DISTRIBUTED BY
operations on AO tables with indexes.

These operations on AO tables with indexes are executed with CT + IIS
whereas on AO tables without indexes which are executed with CTAS

The bug about not preserving storage options is also present
for ALTER TABLE on AO tables without indexes case.
This change fixes that bug by adding the storage options preservation
logic to AT calls for AO tables which go through an internal CTAS.

Co-authored-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>