Fix the test failure on release pipeline
Regress tests failed on the release
pipeline due to the diskquota.naptime
is not 0. As diskquota launcher
firstly sleep, then starts diskquota
workers. When naptime is not 0, workers
are started very late, then some tests
which checking them failed.

In this pr, fix the test by starting
diskquota workers in the launcher before
it sleeps. This makes diskquota workers
will be started immediately when the 
launcher starts.

Also, increased the sleep time in
the tests.

Sometimes, after restarting gpdb cluster, the db
has not yet been added to monitored_db_cache,
but worker_get_epoch is called to fetch the db
from the cache, then it will fail. Shouldn't to
report error here, printing a warning is ok.