brin: Restore specific range summarization - heap
This commit restores specific heap range summarization (range !=
BRIN_ALL_BLOCKRANGES) for heap tables, which is an API used by
autovacuum. Keeping it disabled for AO/CO for now.

There is no reason to have it disabled, except perhaps we need to think
about MPPizing it properly - which is out of scope here.

Also, we want to be able to prevent regressions for heap tables with any
changes we make to BRIN code downstream.

While fixing the tests accordingly, the following improvements have been
made:

* Clearer comments in the tests, specially for summarization.
* Made most tables localized to one QE - it is extremely hard to reason
  about correctness otherwise (since we are dealing with ctids etc)

remove diff