brin ao/co: Ensure final range summarization: build
We weren't "processing the final batch" when we were switching block
sequences (#15155). This led to us not summarizing the final range in
the current block sequence, before switching to the next sequence,
during builds.

Now, we do so by maintaining a counter of the number of tuples processed
for each block sequence, checking that counter and then calling
form_and_insert_tuple(), before switching to the next sequence.

This is a necessary evil, until we refactor brinbuild to loop over
BlockSequences.